Try client test

+++ demo_cerra/src/Client/JavaProgram.java
@@ -4,9 +4,10 @@ import java.util.Vector;
 
 public class JavaProgram implements Job {
 	int id = 0;
+	
 	public Object run(Object parameters) {
 		
-		return new Integer(0);
+		return new Integer(99);
 	}
 
 	@Override
